common: misc improvement and cleanup
    - remove the redundant code
    - move util function together
    - refine the blkid command
    - improve the logger and comment
