+++ pkg/block/blkid.go
@@ -7,60 +7,44 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-func GetFileSystemType(part string) string {
-	if !strings.HasPrefix(part, "/dev") {
-		part = "/dev/" + part
+func doCommandBlkid(partition string, param string) ([]byte, error) {
+	if !strings.HasPrefix(partition, "/dev") {
+		partition = "/dev/" + partition
 	}
 	args := []string{
 		"blkid",
 		"-s",
-		FsType,
-		part,
+		param,
+		partition,
+		"-o",
+		"value",
 	}
-	out, err := exec.Command(args[0], args[1:]...).Output()
+	return exec.Command(args[0], args[1:]...).Output()
+}
+
+func GetFileSystemType(part string) string {
+	out, err := doCommandBlkid(part, FsType)
+
 	if err != nil {
 		logrus.Debugf("failed to read disk uuid of %s : %s\n", part, err.Error())
 		return ""
 	}
 
-	if out == nil || len(out) == 0 {
+	if len(out) == 0 {
 		return ""
 	}
-
-	parts := strings.Split(string(out), "TYPE=")
-	if len(parts) != 2 {
-		logrus.Debugf("failed to parse the type of %s\n", part)
-		return ""
-	}
-
-	return strings.ReplaceAll(strings.TrimSpace(parts[1]), `"`, "")
+	return strings.Split(string(out), "\n")[0]
 }
 
 func GetDiskUUID(part string, uuidType string) string {
-	if !strings.HasPrefix(part, "/dev") {
-		part = "/dev/" + part
-	}
-	args := []string{
-		"blkid",
-		"-s",
-		uuidType,
-		part,
-	}
-	out, err := exec.Command(args[0], args[1:]...).Output()
+	out, err := doCommandBlkid(part, uuidType)
 	if err != nil {
 		logrus.Debugf("failed to read disk uuid of %s : %s\n", part, err.Error())
 		return ""
 	}
 
-	if out == nil || len(out) == 0 {
+	if len(out) == 0 {
 		return ""
 	}
-
-	parts := strings.Split(string(out), "UUID=")
-	if len(parts) != 2 {
-		logrus.Debugf("failed to parse the uuid of %s\n", part)
-		return ""
-	}
-
-	return strings.ReplaceAll(strings.TrimSpace(parts[1]), `"`, "")
+	return strings.Split(string(out), "\n")[0]
 }
