+++ pkg/controller/blockdevice/controller.go
@@ -21,7 +21,6 @@ import (
 
 	diskv1 "github.com/harvester/node-disk-manager/pkg/apis/harvesterhci.io/v1beta1"
 	"github.com/harvester/node-disk-manager/pkg/block"
-	"github.com/harvester/node-disk-manager/pkg/disk"
 	ctldiskv1 "github.com/harvester/node-disk-manager/pkg/generated/controllers/harvesterhci.io/v1beta1"
 	ctllonghornv1 "github.com/harvester/node-disk-manager/pkg/generated/controllers/longhorn.io/v1beta1"
 	"github.com/harvester/node-disk-manager/pkg/option"
@@ -143,7 +142,8 @@ func (c *Controller) OnBlockDeviceChange(key string, device *diskv1.BlockDevice)
 		return nil, fmt.Errorf("failed to resolve persistent dev path for block device %s", device.Name)
 	}
 	filesystem := c.BlockInfo.GetFileSystemInfoByDevPath(devPath)
-	logrus.Debugf("Get filesystem info from device %s, fs type: %s", devPath, filesystem.Type)
+	devPathStatus := convertFSInfoToString(filesystem)
+	logrus.Debugf("Get filesystem info from device %s, %s", devPath, devPathStatus)
 
 	needFormat := deviceCpy.Spec.FileSystem.ForceFormatted && deviceCpy.Status.DeviceStatus.FileSystem.LastFormattedAt == nil
 	if needFormat {
@@ -235,7 +235,7 @@ func (c *Controller) updateDeviceMount(device *diskv1.BlockDevice, devPath strin
 	}
 	if needMountUpdate.Has(NeedMountUpdateUnmount) {
 		logrus.Infof("Unmount device %s from path %s", device.Name, filesystem.MountPoint)
-		if err := disk.UmountDisk(filesystem.MountPoint); err != nil {
+		if err := util.UmountDisk(filesystem.MountPoint); err != nil {
 			return err
 		}
 		diskv1.DeviceMounted.SetError(device, "", nil)
@@ -244,16 +244,13 @@ func (c *Controller) updateDeviceMount(device *diskv1.BlockDevice, devPath strin
 	if needMountUpdate.Has(NeedMountUpdateMount) {
 		expectedMountPoint := extraDiskMountPoint(device)
 		logrus.Infof("Mount deivce %s to %s", device.Name, expectedMountPoint)
-		if err := disk.MountDisk(devPath, expectedMountPoint); err != nil {
+		if err := util.MountDisk(devPath, expectedMountPoint); err != nil {
 			return err
 		}
 		diskv1.DeviceMounted.SetError(device, "", nil)
 		diskv1.DeviceMounted.SetStatusBool(device, true)
 	}
-	if needMountUpdate != NeedMountUpdateNoOp {
-		return c.updateDeviceFileSystem(device, devPath)
-	}
-	return nil
+	return c.updateDeviceFileSystem(device, devPath)
 }
 
 func (c *Controller) updateDeviceFileSystem(device *diskv1.BlockDevice, devPath string) error {
@@ -291,7 +288,7 @@ func (c *Controller) forceFormat(device *diskv1.BlockDevice, devPath string, fil
 	// umount the disk if it is mounted
 	if filesystem != nil && filesystem.MountPoint != "" {
 		logrus.Infof("unmount %s for %s", filesystem.MountPoint, device.Name)
-		if err := disk.UmountDisk(filesystem.MountPoint); err != nil {
+		if err := util.UmountDisk(filesystem.MountPoint); err != nil {
 			return err
 		}
 	}
@@ -319,7 +316,7 @@ func (c *Controller) forceFormat(device *diskv1.BlockDevice, devPath string, fil
 			uuid = ""
 		}
 	}
-	if err := disk.MakeExt4DiskFormatting(devPath, uuid); err != nil {
+	if err := util.MakeExt4DiskFormatting(devPath, uuid); err != nil {
 		return err
 	}
 
@@ -544,7 +541,7 @@ func (c *Controller) OnBlockDeviceDelete(key string, device *diskv1.BlockDevice)
 		}
 		existingMount := bd.Status.DeviceStatus.FileSystem.MountPoint
 		if existingMount != "" {
-			if err := disk.UmountDisk(existingMount); err != nil {
+			if err := util.UmountDisk(existingMount); err != nil {
 				logrus.Warnf("cannot umount disk %s from mount point %s, err: %s", bd.Name, existingMount, err.Error())
 			}
 		}
@@ -652,4 +649,12 @@ func convertMountStr(mountOP NeedMountUpdateOP) string {
 		return "Unmount"
 	}
 	return "Unknown OP"
+}
+
+func convertFSInfoToString(fsInfo *block.FileSystemInfo) string {
+	// means this device is not mounted
+	if fsInfo.MountPoint == "" {
+		return "device is not mounted"
+	}
+	return fmt.Sprintf("mountpoint: %s, fsType: %s", fsInfo.MountPoint, fsInfo.Type)
 }
